Improving memory allocations
Firstly, instead of allocating arrays of variable length on the stack,
we allocated them on the heap. We are using Rcpp's types to do this
allocation, which should use R's memory pool for this.

Secondly, we are avoiding an extra memory allocation and copy operation
by passing down the pixel_matrix's underlying pointer directly as the
output parameter to wcss2p.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>