Remove misleading Additional_repositories
This declaration only generates warnings, as there is actually no R
package repository under that location (there are *git* repositories
though). On the other hand the declaration gives the false sense of
actually indicating the system that dependencies could be fetched
automatically from the git repos under that location, which isn't the
case.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>