Simplify management of wcslib subfolder
On the one hand, we used to call "make all" when invoking make in the
wcslib subfolder, but in reality we only need the static library, not
the shared one.

Secondly, we already have a thorough cleanup script in our top-level
directory that cleans up all the residuals from the wcslib compilation,
so there's no need for us to go and do any further cleanup when "make
clean" is invoked.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>