Treat shared library generically and thoroughly
The shared library produced by wcslib (which we don't actually need, but
existing development installations might have one already in the
filesystem) shouldn't end up in the tarball produced by a "build" R
command, and should also be part of the cleanup process.

The first item was covered already, but for older versions of wcslib. A
while ago we updated to wcslib 7.1 but forgot to update the
.Rbuildignore file to reflect this change. Regarding cleanup, we didn't
include the shared library in the cleanup process, but we should have to
be thorough about it.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>