Temporarily patch wcslib sources to avoid LTO clashes
The prj.c module in wcslib declares a number of global constant
integers, among them "CAR". On the other hand, R also has a function
called "CAR" (declared in Rinternals.h, and defined in Rinlinedfuns.h)
declaration for the same name. This clash of names isn't usually a
problem, because wcslib's compilation doesn't include Rinternals.h (so
there is no clash a object compilation time), and by the time Rwcs.so is
loaded all references to CAR, an internal R function, should have been
resolved already. Note that this should be true even if Rwcs.so is
loaded with RTLD_GLOBAL, because it should be a symbol used only by R
libraries internally, and not by third-party extensions.

This clash *does* cause issues though when building with LTO enabled. In
that case, the link-time optimisation pass sees the two different
definitions at the same time and emits an error.

While there are several ways of solving this issue, the easiest seems to
be to declare wcslib's CAR (et al.) as static, since it doesn't seem to
leave the module where it's defined (and it's not declared in any header
file, further supporting the idea that it's an internal variable). I've
tested this change on top of the latest wcslib release and it works, so
I've asked Mark Callabreta to consider it into the next release. Once
that's done, we'll update our copy of wcslib and remove this temporary
patch.

This was originally reported in #10.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>